requirements.txt
asgiref==3.2.10
Django==3.1
django-filter==2.2.0
djangorestframework==3.11.1
environ==1.0
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
